--- application/app-delivery/src/main/java/nightmarket/appdelivery/AppDeliveryApplication.java
@@ -1,13 +1,0 @@
-package nightmarket.appdelivery;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AppDeliveryApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AppDeliveryApplication.class, args);
-	}
-
-}
