+++ application/app-delivery/src/main/java/store/nightmarket/application/appdelivery/AppDeliveryApplication.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.appdelivery;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AppDeliveryApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AppDeliveryApplication.class, args);
+	}
+
+}
