--- application/app-delivery/src/test/java/nightmarket/appdelivery/AppDeliveryApplicationTests.java
@@ -1,13 +1,0 @@
-package nightmarket.appdelivery;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppDeliveryApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
