+++ application/app-delivery/src/test/java/store/nightmarket/application/appdelivery/AppDeliveryApplicationTests.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.appdelivery;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AppDeliveryApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
