--- application/app-item/src/main/java/nightmarket/item/AppItemApplication.java
@@ -1,13 +1,0 @@
-package nightmarket.item;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AppItemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AppItemApplication.class, args);
-	}
-
-}
