+++ application/app-item/src/main/java/store/nightmarket/application/appitem/AppItemApplication.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.appitem;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AppItemApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AppItemApplication.class, args);
+	}
+
+}
