--- application/app-item/src/test/java/nightmarket/item/AppItemApplicationTests.java
@@ -1,13 +1,0 @@
-package nightmarket.item;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppItemApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
