+++ application/app-item/src/test/java/store/nightmarket/application/appitem/AppItemApplicationTests.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.appitem;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AppItemApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
