--- application/app-order/src/main/java/nightmarket/apporder/AppOrderApplication.java
@@ -1,13 +1,0 @@
-package nightmarket.apporder;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AppOrderApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AppOrderApplication.class, args);
-	}
-
-}
