+++ application/app-order/src/main/java/store/nightmarket/application/apporder/AppOrderApplication.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.apporder;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AppOrderApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AppOrderApplication.class, args);
+	}
+
+}
