--- application/app-order/src/test/java/nightmarket/apporder/AppOrderApplicationTests.java
@@ -1,13 +1,0 @@
-package nightmarket.apporder;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppOrderApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
