+++ application/app-order/src/test/java/store/nightmarket/application/apporder/AppOrderApplicationTests.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.apporder;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AppOrderApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
