--- application/app-payment/src/main/java/nightmarket/apppayment/AppPaymentApplication.java
@@ -1,13 +1,0 @@
-package nightmarket.apppayment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AppPaymentApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AppPaymentApplication.class, args);
-	}
-
-}
