+++ application/app-payment/src/main/java/store/nightmarket/application/apppayment/AppPaymentApplication.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.apppayment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AppPaymentApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AppPaymentApplication.class, args);
+	}
+
+}
