--- application/app-payment/src/test/java/nightmarket/apppayment/AppPaymentApplicationTests.java
@@ -1,13 +1,0 @@
-package nightmarket.apppayment;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppPaymentApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
