+++ application/app-payment/src/test/java/store/nightmarket/application/apppayment/AppPaymentApplicationTests.java
@@ -1,0 +1,13 @@
+package store.nightmarket.application.apppayment;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AppPaymentApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
