--- domain/user/src/main/java/nightmarket/user/UserApplication.java
@@ -1,13 +1,0 @@
-package nightmarket.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserApplication.class, args);
-	}
-
-}
