+++ domain/user/src/main/java/store/nightmarket/domain/user/UserApplication.java
@@ -1,0 +1,13 @@
+package store.nightmarket.domain.user;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserApplication.class, args);
+	}
+
+}
